refactor: rename utils to api
add docs to api
